refactor: 중복되는 forward 로직을 함수형 인터페이스로 분리
- RequestHandler를 구성한 AbstractRequestHandler를 두어서, 함수형 인터페이스를 구현하게 변경
- RequestHandler의 구현체들은 AbstractRequestHandler를 상속받아 기본동작을 그대로 유지하게 함
