--- src/main/java/codesquad/web/handler/RegistrationForwardHandler.java
@@ -1,14 +1,0 @@
-package codesquad.web.handler;
-
-import codesquad.was.RequestHandler;
-import codesquad.was.http.HttpRequest;
-import codesquad.was.http.HttpResponse;
-import java.io.IOException;
-
-public class RegistrationForwardHandler implements RequestHandler {
-
-    @Override
-    public void process(final HttpRequest request, final HttpResponse response) throws IOException {
-        response.forward("/registration/index.html");
-    }
-}
