+++ src/main/java/codesquad/web/handler/SignUpRequestHandler.java
@@ -1,21 +1,21 @@
 package codesquad.web.handler;
 
 import codesquad.model.User;
-import codesquad.was.RequestHandler;
+import codesquad.was.AbstractRequestHandler;
 import codesquad.was.http.HttpRequest;
 import codesquad.was.http.HttpResponse;
-import codesquad.web.io.InMemoryDataBase;
+import codesquad.web.io.InMemoryUserDataBase;
 import java.io.IOException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class SignUpRequestHandler implements RequestHandler {
+public class SignUpRequestHandler extends AbstractRequestHandler {
 
     private final Logger log = LoggerFactory.getLogger(getClass());
 
     @Override
     public void handlePost(final HttpRequest request, final HttpResponse response) throws IOException {
-        long id = InMemoryDataBase.generateId();
+        long id = InMemoryUserDataBase.generateId();
         String userId = request.getParameter("userId");
         String nickname = request.getParameter("nickname");
         String password = request.getParameter("password");
@@ -24,7 +24,7 @@ public class SignUpRequestHandler implements RequestHandler {
         User user = new User(id, userId, nickname, password, email);
         log.info("회원가입 유저 = {}", user);
 
-        InMemoryDataBase.saveUser(user);
-        response.sendRedirect("/main");
+        InMemoryUserDataBase.saveUser(user);
+        response.sendRedirect("/");
     }
 }
